+++ rotkehlchen/tests/unit/decoders/test_gitcoin_v2.py
@@ -11,20 +11,17 @@ from rotkehlchen.tests.utils.ethereum import get_decoded_events_of_transaction
 from rotkehlchen.types import Location, TimestampMS, deserialize_evm_tx_hash
 
 
-@pytest.mark.vcr
+@pytest.mark.vcr(filter_query_parameters=['apikey'])
 @pytest.mark.parametrize('optimism_accounts', [['0x9531C059098e3d194fF87FebB587aB07B30B1306']])
 def test_optimism_donation_received(optimism_inquirer, optimism_accounts):
-    tx_hex = deserialize_evm_tx_hash('0x08685669305ee26060a5a78ae70065aec76d9e62a35f0837c291fb1232f33601')  # noqa: E501
-    evmhash = deserialize_evm_tx_hash(tx_hex)
-    user_address = optimism_accounts[0]
+    tx_hash = deserialize_evm_tx_hash('0x08685669305ee26060a5a78ae70065aec76d9e62a35f0837c291fb1232f33601')  # noqa: E501
     events, _ = get_decoded_events_of_transaction(
         evm_inquirer=optimism_inquirer,
-        tx_hash=tx_hex,
+        tx_hash=tx_hash,
     )
-    timestamp = TimestampMS(1692176477000)
-    amount_str = '0.00122'
+    user_address, timestamp, amount_str = optimism_accounts[0], TimestampMS(1692176477000), '0.00122'  # noqa: E501
     expected_events = [EvmEvent(
-        tx_hash=evmhash,
+        tx_hash=tx_hash,
         sequence_index=0,
         timestamp=timestamp,
         location=Location.OPTIMISM,
@@ -40,15 +37,14 @@ def test_optimism_donation_received(optimism_inquirer, optimism_accounts):
     assert events == expected_events
 
 
-@pytest.mark.vcr
+@pytest.mark.vcr(filter_query_parameters=['apikey'])
 @pytest.mark.parametrize('ethereum_accounts', [['0x9531C059098e3d194fF87FebB587aB07B30B1306']])
 def test_ethereum_donation_received(ethereum_inquirer, ethereum_accounts):
-    tx_hex = deserialize_evm_tx_hash('0x71fc406467f342f5801560a326aa29ac424381daf17cc04b5573960425ba605b')  # noqa: E501
-    evmhash = deserialize_evm_tx_hash(tx_hex)
-    events, _ = get_decoded_events_of_transaction(evm_inquirer=ethereum_inquirer, tx_hash=tx_hex)
+    tx_hash = deserialize_evm_tx_hash('0x71fc406467f342f5801560a326aa29ac424381daf17cc04b5573960425ba605b')  # noqa: E501
+    events, _ = get_decoded_events_of_transaction(evm_inquirer=ethereum_inquirer, tx_hash=tx_hash)
     amount_str = '0.001'
     expected_events = [EvmEvent(
-        tx_hash=evmhash,
+        tx_hash=tx_hash,
         sequence_index=0,
         timestamp=TimestampMS(1683655379000),
         location=Location.ETHEREUM,
@@ -64,23 +60,18 @@ def test_ethereum_donation_received(ethereum_inquirer, ethereum_accounts):
     assert events == expected_events
 
 
-@pytest.mark.vcr
+@pytest.mark.vcr(filter_query_parameters=['apikey'])
 @pytest.mark.parametrize('ethereum_accounts', [[
     '0x1B274eaCc333F4a904D72b576B55A108532aB379',
     # also track a grantee to see we handle donating to self fine
     '0xB352bB4E2A4f27683435f153A259f1B207218b1b',
 ]])
 def test_ethereum_make_donation(ethereum_inquirer, ethereum_accounts):
-    tx_hex = deserialize_evm_tx_hash('0xd8d55b66f19a6dbf260d171fbb0c4c146f00c90919f1215cf691d7f0684771c6')  # noqa: E501
-    evmhash = deserialize_evm_tx_hash(tx_hex)
-    user_address = ethereum_accounts[0]
-    tracked_grant = ethereum_accounts[1]
-    events, _ = get_decoded_events_of_transaction(evm_inquirer=ethereum_inquirer, tx_hash=tx_hex)
-    timestamp = TimestampMS(1683676595000)
-    amount_str = '0.0006'
-    gas_str = '0.011086829409239852'
+    tx_hash = deserialize_evm_tx_hash('0xd8d55b66f19a6dbf260d171fbb0c4c146f00c90919f1215cf691d7f0684771c6')  # noqa: E501
+    events, _ = get_decoded_events_of_transaction(evm_inquirer=ethereum_inquirer, tx_hash=tx_hash)
+    user_address, tracked_grant, timestamp, amount_str, gas_str = ethereum_accounts[0], ethereum_accounts[1], TimestampMS(1683676595000), '0.0006', '0.011086829409239852'  # noqa: E501
     expected_events = [EvmEvent(
-        tx_hash=evmhash,
+        tx_hash=tx_hash,
         sequence_index=0,
         timestamp=timestamp,
         location=Location.ETHEREUM,
@@ -92,7 +83,7 @@ def test_ethereum_make_donation(ethereum_inquirer, ethereum_accounts):
         notes=f'Burned {gas_str} ETH for gas',
         counterparty=CPT_GAS,
     ), EvmEvent(
-        tx_hash=evmhash,
+        tx_hash=tx_hash,
         sequence_index=1,
         timestamp=timestamp,
         location=Location.ETHEREUM,
@@ -107,7 +98,7 @@ def test_ethereum_make_donation(ethereum_inquirer, ethereum_accounts):
     )]
 
     expected_events += [EvmEvent(
-        tx_hash=evmhash,
+        tx_hash=tx_hash,
         sequence_index=idx,
         timestamp=timestamp,
         location=Location.ETHEREUM,
@@ -128,17 +119,14 @@ def test_ethereum_make_donation(ethereum_inquirer, ethereum_accounts):
     assert events == expected_events
 
 
-@pytest.mark.vcr
+@pytest.mark.vcr(filter_query_parameters=['apikey'])
 @pytest.mark.parametrize('optimism_accounts', [['0x9531C059098e3d194fF87FebB587aB07B30B1306']])
 def test_optimism_create_project(optimism_inquirer, optimism_accounts):
-    tx_hex = deserialize_evm_tx_hash('0xe59f04c693e91f1659bd8bc718c993158efeb9af02c9c6337f039c44d8a822f6')  # noqa: E501
-    evmhash = deserialize_evm_tx_hash(tx_hex)
-    user_address = optimism_accounts[0]
-    events, _ = get_decoded_events_of_transaction(evm_inquirer=optimism_inquirer, tx_hash=tx_hex)
-    timestamp = TimestampMS(1691697693000)
-    gas_str = '0.000085459641651569'
+    tx_hash = deserialize_evm_tx_hash('0xe59f04c693e91f1659bd8bc718c993158efeb9af02c9c6337f039c44d8a822f6')  # noqa: E501
+    events, _ = get_decoded_events_of_transaction(evm_inquirer=optimism_inquirer, tx_hash=tx_hash)
+    user_address, timestamp, gas_str = optimism_accounts[0], TimestampMS(1691697693000), '0.000085459641651569'  # noqa: E501
     expected_events = [EvmEvent(
-        tx_hash=evmhash,
+        tx_hash=tx_hash,
         sequence_index=0,
         timestamp=timestamp,
         location=Location.OPTIMISM,
@@ -150,7 +138,7 @@ def test_optimism_create_project(optimism_inquirer, optimism_accounts):
         notes=f'Burned {gas_str} ETH for gas',
         counterparty=CPT_GAS,
     ), EvmEvent(
-        tx_hash=evmhash,
+        tx_hash=tx_hash,
         sequence_index=65,
         timestamp=timestamp,
         location=Location.OPTIMISM,
@@ -163,7 +151,7 @@ def test_optimism_create_project(optimism_inquirer, optimism_accounts):
         counterparty=CPT_GITCOIN,
         address='0x8e1bD5Da87C14dd8e08F7ecc2aBf9D1d558ea174',
     ), EvmEvent(
-        tx_hash=evmhash,
+        tx_hash=tx_hash,
         sequence_index=66,
         timestamp=timestamp,
         location=Location.OPTIMISM,
@@ -179,17 +167,14 @@ def test_optimism_create_project(optimism_inquirer, optimism_accounts):
     assert events == expected_events
 
 
-@pytest.mark.vcr
+@pytest.mark.vcr(filter_query_parameters=['apikey'])
 @pytest.mark.parametrize('ethereum_accounts', [['0x9531C059098e3d194fF87FebB587aB07B30B1306']])
 def test_ethereum_project_apply(ethereum_inquirer, ethereum_accounts):
-    tx_hex = deserialize_evm_tx_hash('0x3e4639d97be450c6d32ce77a146898780b75781caaf004d3c40bae083dec07c7')  # noqa: E501
-    evmhash = deserialize_evm_tx_hash(tx_hex)
-    user_address = ethereum_accounts[0]
-    events, _ = get_decoded_events_of_transaction(evm_inquirer=ethereum_inquirer, tx_hash=tx_hex)
-    timestamp = TimestampMS(1673472803000)
-    gas_str = '0.000645250895735256'
+    tx_hash = deserialize_evm_tx_hash('0x3e4639d97be450c6d32ce77a146898780b75781caaf004d3c40bae083dec07c7')  # noqa: E501
+    events, _ = get_decoded_events_of_transaction(evm_inquirer=ethereum_inquirer, tx_hash=tx_hash)
+    user_address, timestamp, gas_str = ethereum_accounts[0], TimestampMS(1673472803000), '0.000645250895735256'  # noqa: E501
     expected_events = [EvmEvent(
-        tx_hash=evmhash,
+        tx_hash=tx_hash,
         sequence_index=0,
         timestamp=timestamp,
         location=Location.ETHEREUM,
@@ -201,7 +186,7 @@ def test_ethereum_project_apply(ethereum_inquirer, ethereum_accounts):
         notes=f'Burned {gas_str} ETH for gas',
         counterparty=CPT_GAS,
     ), EvmEvent(
-        tx_hash=evmhash,
+        tx_hash=tx_hash,
         sequence_index=413,
         timestamp=timestamp,
         location=Location.ETHEREUM,
@@ -217,17 +202,14 @@ def test_ethereum_project_apply(ethereum_inquirer, ethereum_accounts):
     assert events == expected_events
 
 
-@pytest.mark.vcr
+@pytest.mark.vcr(filter_query_parameters=['apikey'])
 @pytest.mark.parametrize('ethereum_accounts', [['0x9531C059098e3d194fF87FebB587aB07B30B1306']])
 def test_ethereum_project_update(ethereum_inquirer, ethereum_accounts):
-    tx_hex = deserialize_evm_tx_hash('0xfe00fa198eb5395e1d809e017c6b416e882b0aef16e82bf00cd60ce5576bb122')  # noqa: E501
-    evmhash = deserialize_evm_tx_hash(tx_hex)
-    user_address = ethereum_accounts[0]
-    events, _ = get_decoded_events_of_transaction(evm_inquirer=ethereum_inquirer, tx_hash=tx_hex)
-    timestamp = TimestampMS(1681330523000)
-    gas_str = '0.001464795019471285'
+    tx_hash = deserialize_evm_tx_hash('0xfe00fa198eb5395e1d809e017c6b416e882b0aef16e82bf00cd60ce5576bb122')  # noqa: E501
+    events, _ = get_decoded_events_of_transaction(evm_inquirer=ethereum_inquirer, tx_hash=tx_hash)
+    user_address, timestamp, gas_str = ethereum_accounts[0], TimestampMS(1681330523000), '0.001464795019471285'  # noqa: E501
     expected_events = [EvmEvent(
-        tx_hash=evmhash,
+        tx_hash=tx_hash,
         sequence_index=0,
         timestamp=timestamp,
         location=Location.ETHEREUM,
@@ -239,7 +221,7 @@ def test_ethereum_project_update(ethereum_inquirer, ethereum_accounts):
         notes=f'Burned {gas_str} ETH for gas',
         counterparty=CPT_GAS,
     ), EvmEvent(
-        tx_hash=evmhash,
+        tx_hash=tx_hash,
         sequence_index=192,
         timestamp=timestamp,
         location=Location.ETHEREUM,
@@ -255,18 +237,15 @@ def test_ethereum_project_update(ethereum_inquirer, ethereum_accounts):
     assert events == expected_events
 
 
-@pytest.mark.vcr
+@pytest.mark.vcr(filter_query_parameters=['apikey'])
 @pytest.mark.parametrize('optimism_accounts', [['0xd034Fd34eaEe5eC2c413C51936109E12873f4DA5']])
 def test_optimism_many_donations_different_strategies(optimism_inquirer, optimism_accounts):
-    tx_hex = deserialize_evm_tx_hash('0x5d85b436f5f177de6019baa9ecebae285e0def4924546307fac40556bece4cd7')  # noqa: E501
-    evmhash = deserialize_evm_tx_hash(tx_hex)
-    user_address = optimism_accounts[0]
-    events, _ = get_decoded_events_of_transaction(evm_inquirer=optimism_inquirer, tx_hash=tx_hex)
-    timestamp = TimestampMS(1692300843000)
-    gas_str = '0.004506208027331091'
+    tx_hash = deserialize_evm_tx_hash('0x5d85b436f5f177de6019baa9ecebae285e0def4924546307fac40556bece4cd7')  # noqa: E501
+    events, _ = get_decoded_events_of_transaction(evm_inquirer=optimism_inquirer, tx_hash=tx_hash)
+    user_address, timestamp, gas_str = optimism_accounts[0], TimestampMS(1692300843000), '0.004506208027331091'  # noqa: E501
     op_dai = Asset('eip155:10/erc20:0xDA10009cBd5D07dd0CeCc66161FC93D7c9000da1')
     assert events[0:2] == [EvmEvent(
-        tx_hash=evmhash,
+        tx_hash=tx_hash,
         sequence_index=0,
         timestamp=timestamp,
         location=Location.OPTIMISM,
@@ -278,7 +257,7 @@ def test_optimism_many_donations_different_strategies(optimism_inquirer, optimis
         notes=f'Burned {gas_str} ETH for gas',
         counterparty=CPT_GAS,
     ), EvmEvent(
-        tx_hash=evmhash,
+        tx_hash=tx_hash,
         sequence_index=50,
         timestamp=timestamp,
         location=Location.OPTIMISM,
@@ -303,15 +282,14 @@ def test_optimism_many_donations_different_strategies(optimism_inquirer, optimis
         assert event.counterparty == CPT_GITCOIN
 
 
-@pytest.mark.vcr
+@pytest.mark.vcr(filter_query_parameters=['apikey'])
 @pytest.mark.parametrize('optimism_accounts', [['0x9531C059098e3d194fF87FebB587aB07B30B1306']])
 def test_optimism_grant_payout(optimism_inquirer, optimism_accounts):
-    tx_hex = deserialize_evm_tx_hash('0x84110136c94ceb71c72afb27ccb517eb33f77a8a419d125101644e2c43294815')  # noqa: E501
-    evmhash = deserialize_evm_tx_hash(tx_hex)
-    events, _ = get_decoded_events_of_transaction(evm_inquirer=optimism_inquirer, tx_hash=tx_hex)
+    tx_hash = deserialize_evm_tx_hash('0x84110136c94ceb71c72afb27ccb517eb33f77a8a419d125101644e2c43294815')  # noqa: E501
+    events, _ = get_decoded_events_of_transaction(evm_inquirer=optimism_inquirer, tx_hash=tx_hash)
     amount_str = '1228.529999999999934464'
     expected_events = [EvmEvent(
-        tx_hash=evmhash,
+        tx_hash=tx_hash,
         sequence_index=73,
         timestamp=TimestampMS(1696942367000),
         location=Location.OPTIMISM,
@@ -327,15 +305,14 @@ def test_optimism_grant_payout(optimism_inquirer, optimism_accounts):
     assert events == expected_events
 
 
-@pytest.mark.vcr
+@pytest.mark.vcr(filter_query_parameters=['apikey'])
 @pytest.mark.parametrize('ethereum_accounts', [['0x9531C059098e3d194fF87FebB587aB07B30B1306']])
 def test_ethereum_grant_payout(ethereum_inquirer, ethereum_accounts):
-    tx_hex = deserialize_evm_tx_hash('0x66ff5be7841f05cc9cb53fd0307460690f91203c52490f5bbfdeabe8462be50b')  # noqa: E501
-    evmhash = deserialize_evm_tx_hash(tx_hex)
-    events, _ = get_decoded_events_of_transaction(evm_inquirer=ethereum_inquirer, tx_hash=tx_hex)
+    tx_hash = deserialize_evm_tx_hash('0x66ff5be7841f05cc9cb53fd0307460690f91203c52490f5bbfdeabe8462be50b')  # noqa: E501
+    events, _ = get_decoded_events_of_transaction(evm_inquirer=ethereum_inquirer, tx_hash=tx_hash)
     amount_str = '20000'
     expected_events = [EvmEvent(
-        tx_hash=evmhash,
+        tx_hash=tx_hash,
         sequence_index=313,
         timestamp=TimestampMS(1689038123000),
         location=Location.ETHEREUM,
@@ -351,20 +328,17 @@ def test_ethereum_grant_payout(ethereum_inquirer, ethereum_accounts):
     assert events == expected_events
 
 
-@pytest.mark.vcr
+@pytest.mark.vcr(filter_query_parameters=['apikey'])
 @pytest.mark.parametrize('polygon_pos_accounts', [['0x9531C059098e3d194fF87FebB587aB07B30B1306']])
 def test_polygon_donation_matic_received(polygon_pos_inquirer, polygon_pos_accounts):
-    tx_hex = deserialize_evm_tx_hash('0x32837e03ac3e9066f09c1ee0807c533aa1bef5e3119b98dcacdd1ca631bc7ca6')  # noqa: E501
-    evmhash = deserialize_evm_tx_hash(tx_hex)
-    user_address = polygon_pos_accounts[0]
+    tx_hash = deserialize_evm_tx_hash('0x32837e03ac3e9066f09c1ee0807c533aa1bef5e3119b98dcacdd1ca631bc7ca6')  # noqa: E501
     events, _ = get_decoded_events_of_transaction(
         evm_inquirer=polygon_pos_inquirer,
-        tx_hash=tx_hex,
+        tx_hash=tx_hash,
     )
-    timestamp = TimestampMS(1700595622000)
-    amount_str = '4'
+    user_address, timestamp, amount_str = polygon_pos_accounts[0], TimestampMS(1700595622000), '4'
     expected_events = [EvmEvent(
-        tx_hash=evmhash,
+        tx_hash=tx_hash,
         sequence_index=0,
         timestamp=timestamp,
         location=Location.POLYGON_POS,
@@ -380,20 +354,17 @@ def test_polygon_donation_matic_received(polygon_pos_inquirer, polygon_pos_accou
     assert events == expected_events
 
 
-@pytest.mark.vcr
+@pytest.mark.vcr(filter_query_parameters=['apikey'])
 @pytest.mark.parametrize('polygon_pos_accounts', [['0x9531C059098e3d194fF87FebB587aB07B30B1306']])
 def test_polygon_donation_token_received(polygon_pos_inquirer, polygon_pos_accounts):
-    tx_hex = deserialize_evm_tx_hash('0x17601356467af0cfcf3a62f93879b504695b8690545b2b8669da5ec0f3a2a91b')  # noqa: E501
-    evmhash = deserialize_evm_tx_hash(tx_hex)
-    user_address = polygon_pos_accounts[0]
+    tx_hash = deserialize_evm_tx_hash('0x17601356467af0cfcf3a62f93879b504695b8690545b2b8669da5ec0f3a2a91b')  # noqa: E501
     events, _ = get_decoded_events_of_transaction(
         evm_inquirer=polygon_pos_inquirer,
-        tx_hash=tx_hex,
+        tx_hash=tx_hash,
     )
-    timestamp = TimestampMS(1700593336000)
-    amount_str = '1.5'
+    user_address, timestamp, amount_str = polygon_pos_accounts[0], TimestampMS(1700593336000), '1.5'  # noqa: E501
     expected_events = [EvmEvent(
-        tx_hash=evmhash,
+        tx_hash=tx_hash,
         sequence_index=236,
         timestamp=timestamp,
         location=Location.POLYGON_POS,
@@ -409,20 +380,17 @@ def test_polygon_donation_token_received(polygon_pos_inquirer, polygon_pos_accou
     assert events == expected_events
 
 
-@pytest.mark.vcr
+@pytest.mark.vcr(filter_query_parameters=['apikey'])
 @pytest.mark.parametrize('polygon_pos_accounts', [['0x9531C059098e3d194fF87FebB587aB07B30B1306']])
 def test_polygon_apply_to_round(polygon_pos_inquirer, polygon_pos_accounts):
-    tx_hex = deserialize_evm_tx_hash('0x51c1909ce9268f453d4b7136b0fecb72d8da567f406c37014dd8ad8ed05c9a1f')  # noqa: E501
-    evmhash = deserialize_evm_tx_hash(tx_hex)
-    user_address = polygon_pos_accounts[0]
+    tx_hash = deserialize_evm_tx_hash('0x51c1909ce9268f453d4b7136b0fecb72d8da567f406c37014dd8ad8ed05c9a1f')  # noqa: E501
     events, _ = get_decoded_events_of_transaction(
         evm_inquirer=polygon_pos_inquirer,
-        tx_hash=tx_hex,
+        tx_hash=tx_hash,
     )
-    timestamp = TimestampMS(1699442294000)
-    gas_str = '0.05638388497968273'
+    user_address, timestamp, gas_str = polygon_pos_accounts[0], TimestampMS(1699442294000), '0.05638388497968273'  # noqa: E501
     expected_events = [EvmEvent(
-        tx_hash=evmhash,
+        tx_hash=tx_hash,
         sequence_index=0,
         timestamp=timestamp,
         location=Location.POLYGON_POS,
@@ -434,7 +402,7 @@ def test_polygon_apply_to_round(polygon_pos_inquirer, polygon_pos_accounts):
         notes=f'Burned {gas_str} POL for gas',
         counterparty=CPT_GAS,
     ), EvmEvent(
-        tx_hash=evmhash,
+        tx_hash=tx_hash,
         sequence_index=1047,
         timestamp=timestamp,
         location=Location.POLYGON_POS,
@@ -446,5 +414,32 @@ def test_polygon_apply_to_round(polygon_pos_inquirer, polygon_pos_accounts):
         notes='Apply to gitcoin round with project application id 0xbb5864fabd76bd8a9d620dd2cfd089a0507135e6e57d12487d4ffd74a4939538',  # noqa: E501
         counterparty=CPT_GITCOIN,
         address='0xa1D52F9b5339792651861329A046dD912761E9A9',
+    )]
+    assert events == expected_events
+
+
+@pytest.mark.vcr(filter_query_parameters=['apikey'])
+@pytest.mark.parametrize('ethereum_accounts', [['0x9531C059098e3d194fF87FebB587aB07B30B1306']])
+def test_ethereum_voted_without_application_index(ethereum_inquirer, ethereum_accounts):
+    """Test that voted events missing the application index are properly seen as donations"""
+    tx_hash = deserialize_evm_tx_hash('0xdcb6d2282b34a3cb7637ac65d8b7f1d0e8f2bc149379767f0b6f9ba2afa8a359')  # noqa: E501
+    events, _ = get_decoded_events_of_transaction(
+        evm_inquirer=ethereum_inquirer,
+        tx_hash=tx_hash,
+    )
+    user_address, timestamp, amount = ethereum_accounts[0], TimestampMS(1673960015000), '0.0035'
+    expected_events = [EvmEvent(
+        tx_hash=tx_hash,
+        sequence_index=0,
+        timestamp=timestamp,
+        location=Location.ETHEREUM,
+        event_type=HistoryEventType.RECEIVE,
+        event_subtype=HistoryEventSubType.DONATE,
+        asset=A_ETH,
+        balance=Balance(amount=FVal(amount)),
+        location_label=user_address,
+        notes=f'Receive a gitcoin donation of {amount} ETH from 0xcD9a4e7C2ad6AAae7Ac25c2139d71739d9Fa2284',  # noqa: E501
+        counterparty=CPT_GITCOIN,
+        address='0x8fBEa07446DdF4518b1a7BA2B4f11Bd140a8DF41',
     )]
     assert events == expected_events